+++ src/Surfnet/StepupGateway/ApiBundle/DependencyInjection/Compiler/SmsAdapterProviderPass.php
@@ -1,0 +1,38 @@
+<?php
+
+/**
+ * Copyright 2021 SURFnet B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Surfnet\StepupGateway\ApiBundle\DependencyInjection\Compiler;
+
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
+
+/**
+ * @SuppressWarnings(PHPMD.UnusedLocalVariables)
+ */
+class SmsAdapterProviderPass implements CompilerPassInterface
+{
+    public function process(ContainerBuilder $container): void
+    {
+        $definition = $container->findDefinition('surfnet_gateway_api.service.provider');
+        $taggedServices = $container->findTaggedServiceIds('sms_provider');
+        foreach ($taggedServices as $id => $tags) {
+            $definition->addMethodCall('addSmsAdapter', [new Reference($id)]);
+        }
+    }
+}