+++ src/Surfnet/StepupGateway/U2fVerificationBundle/Console/Command/MigrationsDiffDoctrineCommand.php
@@ -24,9 +24,6 @@ use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-/**
- * @SuppressWarnings(PHPMD.UnusedLocalVariables)
- */
 class MigrationsDiffDoctrineCommand extends Command
 {
     /**