+++ src/Surfnet/StepupGateway/U2fVerificationBundle/Tests/Service/VerificationServiceRegistrationTest.php
@@ -19,6 +19,7 @@
 namespace Surfnet\StepupGateway\U2fVerificationBundle\Tests\Service;
 
 use Mockery as m;
+use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\NullLogger;
 use Surfnet\StepupGateway\U2fVerificationBundle\Service\VerificationService;
@@ -29,6 +30,8 @@ use Surfnet\StepupU2fBundle\Service\RegistrationVerificationResult;
 
 final class VerificationServiceRegistrationTest extends TestCase
 {
+    use MockeryPHPUnitIntegration;
+
     const APP_ID = 'https://gateway.surfconext.invalid/u2f/app-id';
 
     /**